Added main.css and put code into story.js
Stole some css from www.abeautifulsite.net to make my html look slightly
better. Filled in the JS file with some variables to store an adjective,
verb and noun to fill in the story. The JS calls the prompt() method to
grab the strings from the user and fills them in to the story using some
concatonation and the document.write() method to spit out the
concatonated string variables to the html document.